Additional checks and prep for idiomatic processing

+++ src/FLBConverter/FLBConverter.java
@@ -1,3 +1,11 @@
+/**
+ * @author Phil Wilkins
+ * @Email code@mp3monster.org
+ * @license Apache 2.0
+ * 
+ * This is a single file application (meaning we can skip the jar generation task).
+ * Its purpose is to read a Fluent Bit classic file and generate a YAML representation
+ */
 package FLBConvertor;
 
 import java.io.BufferedReader;
@@ -6,6 +14,7 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.io.FileNotFoundException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
@@ -15,11 +24,23 @@ import java.util.Iterator;
 class FLBConverter {
 
   private static final String FLB_PATH_PREFIX = "FLB_PATH_PREFIX";
-  private static final String NL = "\n";
+  private static final String FLB_PATH_PREFIX_HELP = "Allows us to create a default offset location for our files. Helps when the folder in which the files to be converted are in a different folder";
+
   private static final String FLB_REPORT_FILE = "FLB_REPORT_FILE";
-  private static final String TRUE = "true";
+  private static final String FLB_REPORT_FILE_HELP = "When set to true then the log output will be added to a file";
+
   private static final String FLB_CONVERT_DEBUG = "FLB_CONVERT_DEBUG";
+  private static final String FLB_CONVERT_DEBUG_HELP = "when the value of trues is found debug level logging is enabled.";
+
   private static final String CONVERSION_LIST = "conversion.list";
+  private static final String CONVERSION_LIST_HELP = "Conversion list file provides a means by which we can supply multiple files to convert in one a single execution";
+
+  private static final String FLB_CLASSIC_FN = "FLBClassicFN";
+  private static final String FLB_CLASSIC_FN_HELP = "This environment variable can be used to identify a single conversion";
+
+  private static final String NL = "\n";
+  private static final String TRUE = "true";
+  private static final String REPORT_EXTN = ".report";
   private static final String PIPELINEYAMLLBL = "pipeline:\n";
   private static final String SERVICEYAMLLBL = "service:\n";
   private static final String FILTERCLASSIC = "[FILTER]";
@@ -30,13 +51,19 @@ class FLBConverter {
   private static final String FILTERSYAML = "filters";
   private static final String INPUTSYAML = "inputs";
 
-  private static final String FLB_CLASSIC_FN = "FLBClassicFN";
+  private static final String INFO_LBL = "info:";
+  private static final String DEBUG_LBL = "DEBUG:";
+  private static final String ERROR_LBL = "ERROR:";
 
+  /**
+   * Define the debug flag globally - nothing is gained by passing it around
+   */
   private static boolean debug = false;
+  private static boolean useIdiomaticForm = false;
   private static boolean logToFile = false;
   private static FileWriter converterReport = null;
 
-  /*
+  /**
    * Define the different plugin types.
    */
   enum PluginType {
@@ -47,11 +74,14 @@ class FLBConverter {
   }
 
   /**
-   * @param msg
+   * Writes a debug message if debug is allowed. If the report file is setup
+   * then we'll also add the debug message to that as well
+   * 
+   * @param msg the message to write as a dubg statment
    */
   static void debug(String msg) {
     if (debug) {
-      final String logStr = "debug:" + msg;
+      final String logStr = DEBUG_LBL + msg;
       System.out.println(logStr);
       if (logToFile && converterReport != null) {
         try {
@@ -64,10 +94,13 @@ class FLBConverter {
   }
 
   /**
-   * @param msg
+   * Writes a log message. If the report file is setup
+   * then we'll also add the debug message to that as well
+   * 
+   * @param msg message to be logged
    */
   static void info(String msg) {
-    final String logStr = "info:" + msg;
+    final String logStr = INFO_LBL + msg;
     System.out.println(logStr);
     if (logToFile && converterReport != null) {
       try {
@@ -79,10 +112,13 @@ class FLBConverter {
   }
 
   /**
-   * @param msg
+   * If an error needs to be logged use this method. If the file is enabled then
+   * we add the log to the file as well.
+   * 
+   * @param msg message to be displayed
    */
   static void err(String msg) {
-    final String logStr = "ERROR:" + msg;
+    final String logStr = ERROR_LBL + msg;
     System.out.println(logStr);
     if (logToFile && converterReport != null) {
       try {
@@ -95,7 +131,8 @@ class FLBConverter {
 
   /**
    * This class helps us process the plugins and directives. Each plugin will have
-   * an instance of this class
+   * an instance of this class. We track the plugin type so we can get the order
+   * of output correct
    */
   static class Plugin {
     private static final String WILDCARD = "*";
@@ -103,13 +140,19 @@ class FLBConverter {
     private static final String COMMENT = "#";
     private static final String INCLUDE = "@include";
     private static final String NAMEATTR = "name";
-    public PluginType pluginType;
-    private String name = null;
     private static final String SEPARATOR = " ";
     static final int PLUGININDENT = 1;
     static final int NAMEINDENT = 2;
     static final int ATTRIBUTEINDENT = 3;
+    public PluginType pluginType;
+    private String name = null;
 
+    /**
+     * As certin attributes are allowed to reoccur such as the rules in the modifier
+     * filter plugin we store each attribute name as a HashMap containing array
+     * lists. We've declared our our own class for this to simplify the use of the
+     * definition
+     */
     HashMapArray attributes = new HashMapArray();
 
     String indenter(int depth) {
@@ -124,6 +167,15 @@ class FLBConverter {
       this.pluginType = type;
     }
 
+    /**
+     * We examine the line to see if the @includes is used. Fluent Bit allows us to
+     * set attribute elements with an @includes, however in YAML its use is a lot
+     * more restrictive. So when we encounter its use we need to log it
+     * 
+     * @param line   the line to examine
+     * @param lineNo position in the source file
+     * @return true if an includes is identified
+     */
     static boolean checkForInclusion(String line, int lineNo) {
       boolean found = false;
       String tempLine = line.toLowerCase();
@@ -135,7 +187,46 @@ class FLBConverter {
       return found;
     }
 
+    private String toIdiomaticForm(String attributeName) {
+
+      StringBuilder builder = new StringBuilder(attributeName);
+
+      // Traverse the string character by
+      // character and remove underscore
+      // and capitalize next letter
+      for (int i = 0; i < builder.length(); i++) {
+
+        // Check char is underscore
+        if (builder.charAt(i) == '_') {
+
+          builder.deleteCharAt(i);
+          builder.replace(
+              i, i + 1,
+              String.valueOf(
+                  Character.toUpperCase(
+                      builder.charAt(i))));
+        }
+      }
+
+      // Return in String type
+      return builder.toString();
+
+    }
+
+    /**
+     * 
+     * Each attribute is processed into this instance of a plugin.
+     * If any issues are identified such as the use of
+     * the @includes feature this is logged to
+     * info. We also deal with any of the foible such as quoting
+     * wildcards and dummy attributes
+     * 
+     * @param attribute the attribute line from the source file
+     * @param lineNo    the line number for this attribute in the source file
+     * 
+     */
     public void add(String attribute, int lineNo) {
+      boolean hasInclusion = false;
       if (attribute == null) {
         return;
       }
@@ -144,7 +235,7 @@ class FLBConverter {
       if (attribute.length() == 0) {
         return;
       }
-      checkForInclusion(attribute, lineNo);
+      hasInclusion = checkForInclusion(attribute, lineNo);
 
       String attributeName = null;
       int sepPos = 0;
@@ -155,15 +246,22 @@ class FLBConverter {
       }
       if (sepPos > 0) {
         attributeName = attribute.substring(0, sepPos).trim();
+        if (hasInclusion) {
+          attributeName = "#" + attributeName;
+          // its not already a comment - let's comment out the inclusion
+        }
+        // we need to handle the name attribute slightly differently
         if (attributeName.equalsIgnoreCase(NAMEATTR)) {
           this.name = attribute.substring(sepPos);
         } else {
+          // determine whether there is a wildcard involved, correct the quotations
           String attributeValue = attribute.substring(sepPos).trim();
           if (attributeValue.equalsIgnoreCase(WILDCARD)) {
-            attributeValue = "\"*\"";
+            attributeValue = "'*'";
           }
-          if ((attributeName.equalsIgnoreCase(DUMMYATTR)) && (!attributeValue.startsWith("\'"))) {
-            attributeValue = "\'" + attributeValue + "\'";
+          // ensure that dummy attributes are correctly quoted
+          if ((attributeName.equalsIgnoreCase(DUMMYATTR)) && (!attributeValue.startsWith("'"))) {
+            attributeValue = "'" + attributeValue + "'";
           }
           ArrayList<String> values = null;
           if (attributes.containsKey(attributeName)) {
@@ -180,10 +278,24 @@ class FLBConverter {
 
     }
 
+    /**
+     * Provides a correctly formatted plugin name - we declare this separatelty so
+     * anything that is handled using the plugin object such as a service can have
+     * this action overridden
+     * 
+     * @return returns the name label with correct indentation for the YAML config
+     */
     String writePrefix() {
       return indenter(NAMEINDENT) + "- name: " + this.name + NL;
     }
 
+    /**
+     * We build the YAML representation of the plugin, iterating over the data
+     * structure and handling the possibility of having multiple attributes of the
+     * same type
+     * 
+     * @return string containing the plugin in YAML format
+     */
     public String write() {
       String YAMLoutput = writePrefix();
       Iterator<String> iter = attributes.keySet().iterator();
@@ -210,21 +322,33 @@ class FLBConverter {
       return YAMLoutput;
     }
 
-    public String getName() {
-      return pluginType.toString();
-    }
+    // public String getName() {
+    // return pluginType.toString();
+    // }
   }
 
+  /**
+   * We need the rewepresentation of services to differ from inputs, outputs and
+   * filters. The level of YAML indentation is different, and we don't have a name
+   * attribute
+   */
   static class ServicePlugin extends Plugin {
     public ServicePlugin() {
       super(PluginType.SERVICE);
     }
 
+    /**
+     * Provide identation that reflects the service block
+     */
     @Override
     String indenter(int depth) {
       return "  ";
     }
 
+    /**
+     * We don't want to handle a name attribute in a manner that is different to
+     * other service attributes.
+     */
     @Override
     String writePrefix() {
       return "";
@@ -239,6 +363,9 @@ class FLBConverter {
   private static ArrayList<Plugin> filters = null;
 
   /**
+   * Depending upon the the label in the classic file, we need to decide which
+   * group of plugins to add the latest definition to.
+   * 
    * @param currentPlugin
    */
   private static void storePlugin(Plugin currentPlugin) {
@@ -266,6 +393,14 @@ class FLBConverter {
     }
   }
 
+  /**
+   * This method reads the classic configuration file and handles when we need to
+   * start a new plugin object, and when to add a plugin into the correct array of
+   * plugins. Note we only allow one service definition
+   * 
+   * @param classicFile the reader object for the classiv file
+   * @throws IOException if we fail to read the file properly
+   */
   private static void consumeClassicFile(BufferedReader classicFile) throws IOException {
     Plugin currentPlugin = null;
     int lineCount = 0;
@@ -302,6 +437,15 @@ class FLBConverter {
 
   }
 
+  /**
+   * Takes the list odf plugins, and writes the correct prefix label before
+   * iterrating through the plugins adding them in the correct YAML format
+   * 
+   * @param plugins the list of plugin objects of a particular type
+   * @param outFile the output file buffer
+   * @param label   the label for this type of plugin
+   * @throws IOException thrown if there id
+   */
   private static void writePlugins(ArrayList<Plugin> plugins, BufferedWriter outFile, String label) throws IOException {
     if (!plugins.isEmpty()) {
       outFile.write("  " + label + ":\n");
@@ -313,7 +457,14 @@ class FLBConverter {
 
   }
 
-  private static void writeOutput(BufferedWriter outFile) throws IOException {
+  /**
+   * This orchestrates the corret order in which the pipeline is constructed in
+   * the YAML file.
+   * 
+   * @param outFile the buffer write for the output
+   * @throws IOException any io errors should lead to us bailing
+   */
+  private static void writePipelineOutput(BufferedWriter outFile) throws IOException {
     if (service != null) {
       outFile.write(SERVICEYAMLLBL);
       outFile.write(service.write());
@@ -336,6 +487,13 @@ class FLBConverter {
     outFile.close();
   }
 
+  /**
+   * This works out the name for the output file based on the input name
+   * 
+   * @param inFileName input (classic) filename
+   * @return output file's name.
+   * 
+   */
   private static String getOutFile(String inFileName) {
     String outFileName = inFileName;
 
@@ -346,24 +504,37 @@ class FLBConverter {
     return outFileName + "yaml";
   }
 
+  /**
+   * We don't pass the filename of the report file as this is handled by the
+   * logging methods. This method drives the core logic of scanning the Classic
+   * file format and once read runs the process of writing each plugin/directive
+   * to the output file
+   * 
+   * @param inFileName
+   * 
+   * @param outFileName
+   */
   private static void processor(String inFileName, String outFileName) {
     getOutFile(inFileName);
     BufferedWriter outFile = null;
     FileReader fr = null;
     try {
-      fr = new FileReader(inFileName);
+      File inFile = new File(inFileName);
+      if (!inFile.exists()) {
+        err("Cant locate input file:" + inFileName);
+        return;
+      }
+      fr = new FileReader(inFile);
       info("InputFile:" + inFileName + " --> " + outFileName);
       consumeClassicFile(new BufferedReader(fr));
 
       info("Inputs:" + inputs.size() + NL + "Filters:" + filters.size() + NL + "Outputs:" + outputs.size() + NL);
 
       outFile = new BufferedWriter(new FileWriter(outFileName));
-      writeOutput(outFile);
+      writePipelineOutput(outFile);
       fr.close();
 
-    } catch (
-
-    Exception err) {
+    } catch (Exception err) {
       err(err.toString());
       err.printStackTrace();
 
@@ -380,17 +551,26 @@ class FLBConverter {
           fr.close();
         }
       } catch (IOException ioErr) {
-        err("finally fr err closing file reader\n" + ioErr.getMessage());
+        err("ioError fr err closing file reader\n" + ioErr.getMessage());
       }
     }
   }
 
+  /*
+   * Generates the current date and time as a string
+   */
   static String getDateStr() {
     DateTimeFormatter dtf = DateTimeFormatter.ofPattern("dd/MM/yyyy HH:mm:ss");
     LocalDateTime now = LocalDateTime.now();
     return (dtf.format(now));
   }
 
+  /**
+   * Checks for the environment variable that control whether we log to debug
+   * level. We've not adopted a logging framework so that we can keep the code as
+   * a
+   * single class. This means we can run without needing to have a Mave or Gradle
+   */
   private static boolean checkDebug() {
     String debugFlagStr = System.getenv(FLB_CONVERT_DEBUG);
     if ((debugFlagStr != null) && (debugFlagStr.trim().equalsIgnoreCase(TRUE))) {
@@ -400,6 +580,11 @@ class FLBConverter {
     return debug;
   }
 
+  /**
+   * If the environment variable is set then any info logs are also written to a
+   * file that is the same as the output with the filename extended with .report
+   * Useful if we're using a conversion.list
+   */
   private static boolean checkReportToFile() {
     String reportFlagStr = System.getenv(FLB_REPORT_FILE);
     logToFile = false;
@@ -410,12 +595,16 @@ class FLBConverter {
     return logToFile;
   }
 
+  /**
+   * Gets the path prefix env var if set. Then if it is not terminated with slash
+   * we add a slash.
+   */
   private static String getPathPrefix() {
     String pathPrefix = System.getenv(FLB_PATH_PREFIX);
-    pathPrefix = pathPrefix.trim();
     if ((pathPrefix == null) || (pathPrefix.length() == 0)) {
       pathPrefix = "";
     } else {
+      pathPrefix = pathPrefix.trim();
       if ((!pathPrefix.endsWith("\\")) || (!pathPrefix.endsWith("/"))) {
         if (pathPrefix.contains("\\")) {
           pathPrefix = pathPrefix + "\\";
@@ -430,9 +619,39 @@ class FLBConverter {
 
   }
 
+  /**
+   * Checks for the idomatic form configuration and sets the flag
+   * 
+   * @return
+   */
+  private static boolean useIdiomatricForm() {
+    boolean flag = false;
+    String idiomaticFlagStr = System.getenv("FLB_IDIOMATICFORM");
+    logToFile = false;
+    if ((idiomaticFlagStr != null) && (idiomaticFlagStr.trim().equalsIgnoreCase(TRUE))) {
+      flag = true;
+    }
+    debug("Env flag for using idomatic form to file set to " + idiomaticFlagStr);
+
+    useIdiomaticForm = flag;
+    return flag;
+  }
+
+  /**
+   * 
+   * The app's name drives the execution of processing one or more classic format
+   * files getting the inputfile from one of several different sources. It
+   * establishes the data structures that we use to manage the holding of the
+   * different plugins.
+   * Each new file is processed we reset the constructs to ensure there isn't any
+   * accidental cross contamination
+   * 
+   * @param args
+   */
   public static void main(String[] args) {
     info("Fluent Bit Converter starting ...");
     checkDebug();
+    useIdiomatricForm();
     String inFileName = null;
     try {
       boolean more = false;
@@ -468,7 +687,7 @@ class FLBConverter {
         filters = new ArrayList<Plugin>();
 
         if (checkReportToFile()) {
-          File reportFile = new File(getOutFile(inFileName) + ".report");
+          File reportFile = new File(getOutFile(inFileName) + REPORT_EXTN);
           converterReport = new FileWriter(reportFile);
           converterReport.write("Execution date:" + getDateStr() + NL);
         }
@@ -489,5 +708,19 @@ class FLBConverter {
       err(err.getMessage());
       err.printStackTrace();
     }
+  }
+
+  /**
+   * Generates the CLI help
+   */
+  static void printHelp() {
+    System.out.println(FLB_CONVERT_DEBUG + " -- " + FLB_CONVERT_DEBUG_HELP);
+    System.out.println(FLB_REPORT_FILE + " -- " + FLB_REPORT_FILE_HELP);
+    System.out.println(FLB_PATH_PREFIX + " -- " + FLB_PATH_PREFIX_HELP);
+    System.out.println(FLB_CLASSIC_FN + " -- " + FLB_CLASSIC_FN_HELP);
+
+    System.out.println(NL);
+    System.out.println(CONVERSION_LIST + " -- " + CONVERSION_LIST_HELP);
+
   }
 }
